% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_rainevent_stats.R
\name{calculate_rainevent_stats}
\alias{calculate_rainevent_stats}
\title{Calculate Rain Events}
\usage{
calculate_rainevent_stats(
  results_system,
  aggregation_function = "sum",
  signalThreshold = 0,
  eventSeparationTime = 6 * 3600,
  ...
)
}
\arguments{
\item{results_system}{data frame as retrieved by \code{\link{get_results}}}

\item{aggregation_function}{function to be used for aggregation and passed to
\link[kwb.event]{getEventStatistics} (default: "sum")}

\item{signalThreshold}{passed to \link[kwb.event]{getEvents}, Value that needs
to be exceeded (signalComparisonOperator == "gt") or reached (signalComparisonOperator == "ge")
by the rain heights (or intensities) in order to be counted as a "signal".
default: 0}

\item{eventSeparationTime}{eventSeparationTime passed to \link[kwb.event]{getEvents}
(default: 6*3600)}

\item{...}{additional arguments passed to \link[kwb.event]{getEvents}}
}
\value{
tibble with statistics for all rain events
}
\description{
Calculate Rain Events
}
\examples{
\dontrun{
path_out_file <- "path-to-my-swmm-output-file"
results_system <- kwb.swmm::get_results(path_out = path_out_file)
rainevent_stats <- calculate_rainevent_stats(results_system,
                           aggregation_function = "sum",
                           signalThreshold = 0,
                           eventSeparationTime = 6*3600)
}
}
