% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_export.R
\name{conda_export}
\alias{conda_export}
\title{conda_export}
\usage{
conda_export(condaenv, export_dir = tempdir())
}
\arguments{
\item{condaenv}{name or path of conda environment to export}

\item{export_dir}{export directory (default: tempdir())}
}
\value{
creates  export_dir/environment_condaenv.yml and returns path
}
\description{
conda_export
}
