% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_swmm_executable.R
\name{download_swmm_executable}
\alias{download_swmm_executable}
\title{Download SWMM executable (currently only for windows only!)}
\usage{
download_swmm_executable(tdir = tempdir(), swmm_version = "5.2.0")
}
\arguments{
\item{tdir}{target directory for unzipping zip files}

\item{swmm_version}{desired SWMM version, one of: "5.2.0", "5.1.15", "5.1.14",
"5.1.13" (default: "5.2.0")}
}
\value{
path to download "runswmm.exe"
}
\description{
downloads \code{SWMM} executable from USEPA`s GitHub repo of
\url{https://github.com/USEPA/Stormwater-Management-Model/releases}
}
\examples{
download_swmm_executable()

}
