% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lid_para_types.R
\name{get_lid_para_types}
\alias{get_lid_para_types}
\title{get_lid_para_types}
\usage{
get_lid_para_types()
}
\value{
returns tidy data frame with required parameteration for LIDs based
on SWMM documentation
}
\description{
get_lid_para_types
}
\examples{
get_lid_para_types()

}
