% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results_system.R
\name{get_meta}
\alias{get_meta}
\title{Helper function: get metadata for selected elements}
\usage{
get_meta(name, path_out)
}
\arguments{
\item{name}{of element to select. one of: c("subcatchments", "nodes", "links", "system")}

\item{path_out}{path to SWMM output file}
}
\value{
tibble with columns "name", "itype", "vindex" and corresponding IDs required for \link[swmmr]{read_out}
}
\description{
Helper function: get metadata for selected elements
}
\examples{
get_itype(name = "subcatchments")
}
