% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results_system.R
\name{get_results}
\alias{get_results}
\title{Get Results}
\usage{
get_results(path_out, type = "system", object_name = NULL, vIndex = NULL)
}
\arguments{
\item{path_out}{path to SWMM output file}

\item{type}{output type, select from c("subcatchments" "nodes", "links" or "system"),
default: "system"}

\item{object_name}{Sets the objects of which time series data is returned.
if NULL all objects will be returned (default: NULL)}

\item{vIndex}{indexes of system wide indexes to be imported. By default
all indexes will be imported (default: NULL). If only selected vIndex`s should
be imported have a look at the documentation of \link[swmmr]{read_out}}
}
\value{
for "system" a tibble with selected system variables, for all other "type"
a list with sublists for each "object_name"
}
\description{
Get Results
}
\examples{
\dontrun{
path_out_file <- "path-to-my-swmm-output-file"
results_subcatchments <- kwb.swmm::get_results(path_out_file, type = "subcatchments")
results_links <- kwb.swmm::get_results(path_out_file, type = "links")
results_nodes <- kwb.swmm::get_results(path_out_file, type = "nodes")
results_system <- kwb.swmm::get_results(path_out_file, type = "system")
}
}
