% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results_system.R
\name{get_results_system}
\alias{get_results_system}
\title{Get Results for System}
\usage{
get_results_system(path_out, vIndex = 1:14)
}
\arguments{
\item{path_out}{path to SWMM output file}

\item{vIndex}{indexes of system wide indexes to be imported. By default
all indexes will be imported (default: 1:14). If only selected vIndex`s should
be imported have a look at the documentation of \link[swmmr]{read_out}}
}
\value{
tibble with selected system variables
}
\description{
Get Results for System
}
