% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hectar_to_squaremeter}
\alias{hectar_to_squaremeter}
\title{Helper function: hectar to squaremeter}
\usage{
hectar_to_squaremeter(hectar)
}
\arguments{
\item{hectar}{hectar}
}
\value{
area in hectar
}
\description{
Helper function: hectar to squaremeter
}
\examples{
squaremeter_to_hectar(100)
}
