% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lps_to_mmPerHour.R
\name{lps_to_mmPerHour}
\alias{lps_to_mmPerHour}
\title{Helper function: converts l/s into mm/h}
\usage{
lps_to_mmPerHour(values)
}
\arguments{
\item{values}{values in liter per second}
}
\value{
values in mm per hour
}
\description{
Helper function: converts l/s into mm/h
}
