% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_swmm.R
\name{run_swmm}
\alias{run_swmm}
\title{Run SWMM}
\usage{
run_swmm(model_dir, model_inp, exe, ...)
}
\arguments{
\item{model_dir}{model directory with 'model.inp'}

\item{model_inp}{name of model input fike 'model.inp'}

\item{exe}{path to SWMM executabvle}

\item{...}{additional arguments passed to \code{shell}}
}
\value{
runs SWMM
}
\description{
allows to run SWMM model in model folder, i.e. it is possible
to define paths to climate data in swmm_model.inp with relative paths (if these
are contained in the same folder or a subfolder of the \code{model_dir})
}
