% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{squaremeter_to_hectar}
\alias{squaremeter_to_hectar}
\title{Convert squaremeter to hectar}
\usage{
squaremeter_to_hectar(squaremeter)
}
\arguments{
\item{squaremeter}{squaremeter}
}
\value{
returns hectars
}
\description{
Convert squaremeter to hectar
}
\examples{
squaremeter_to_hectar(10000)

}
